All parsed, nothing saved

+++ src/parser/parser.h
@@ -13,7 +13,7 @@ typedef struct s_field_info
 	bool is_array;
 	u_int8_t type;
 	u_int8_t arr_type;
-	size_t array_size;
+	int array_size;
 } t_field_info;
 
 void please_parse_field(cJSON *json_chain, struct s_field_info *field_info, size_t *__data);
